move earnings email functions to app/marketing/common/earnings.py
